+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/action/SearchJqlAction.kt
@@ -1,47 +1,25 @@
 package com.atlassian.performance.tools.jiraactions.api.action
 
 import com.atlassian.performance.tools.jiraactions.api.SEARCH_WITH_JQL
+import com.atlassian.performance.tools.jiraactions.api.SeededRandom
 import com.atlassian.performance.tools.jiraactions.api.WebJira
 import com.atlassian.performance.tools.jiraactions.api.measure.ActionMeter
 import com.atlassian.performance.tools.jiraactions.api.memories.IssueKeyMemory
 import com.atlassian.performance.tools.jiraactions.api.memories.JqlMemory
-import com.atlassian.performance.tools.jiraactions.api.observation.SearchJqlObservation
-import com.atlassian.performance.tools.jiraactions.api.page.IssueNavigatorPage
-import com.atlassian.performance.tools.jiraactions.api.page.issuenav.DetailView
-import com.atlassian.performance.tools.jiraactions.api.page.issuenav.IssueNavResultsView
-import org.apache.logging.log4j.LogManager
-import org.apache.logging.log4j.Logger
-import javax.json.JsonObject
 
+@Deprecated("Use SearchIssues.Builder")
 class SearchJqlAction(
-    private val jira: WebJira,
-    private val meter: ActionMeter,
-    private val jqlMemory: JqlMemory,
-    private val issueKeyMemory: IssueKeyMemory,
-    private val view: IssueNavResultsView = DetailView(jira.driver)
+    jira: WebJira,
+    meter: ActionMeter,
+    jqlMemory: JqlMemory,
+    issueKeyMemory: IssueKeyMemory
 ) : Action {
-    private val logger: Logger = LogManager.getLogger(this::class.java)
 
-    override fun run() {
-        val jqlQuery = jqlMemory.recall()
-        if (jqlQuery == null) {
-            logger.debug("Skipping ${SEARCH_WITH_JQL.label} action. I have no knowledge of JQL queries.")
-            return
-        }
+    private val action = SearchIssues.Builder(jira, meter, SeededRandom())
+        .actionType(SEARCH_WITH_JQL)
+        .jqlMemory(jqlMemory)
+        .issueKeyMemory(issueKeyMemory)
+        .build()
 
-        val issueNavigatorPage = meter.measure(
-            key = SEARCH_WITH_JQL,
-            action = { jira.goToIssueNavigator(jqlQuery).also { view.switchToView() }.waitForResults(view) },
-            observation = this::observe
-        )
-        issueKeyMemory.remember(issueNavigatorPage.getIssueKeys())
-    }
-
-    private fun observe(
-        page: IssueNavigatorPage
-    ): JsonObject {
-        val issueKeys = page.getIssueKeys()
-        issueKeyMemory.remember(issueKeys)
-        return SearchJqlObservation(page.jql, issueKeys.size, page.getTotalResults()).serialize()
-    }
+    override fun run() = action.run()
 }